# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform-wide statistics
      security:
        - adminBearer: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all app users
      security:
        - adminBearer: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a single user with full detail
      security:
        - adminBearer: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "404":
          description: Not found
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Suspend or reactivate a user
      security:
        - adminBearer: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "404":
          description: Not found
  /admin/otp-config:
    get:
      operationId: getAdminOtpConfig
      tags: [admin]
      summary: Get OTP gateway configuration
      security:
        - adminBearer: []
      responses:
        "200":
          description: OTP config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOtpConfig"
    put:
      operationId: updateAdminOtpConfig
      tags: [admin]
      summary: Update OTP gateway configuration
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOtpConfigInput"
      responses:
        "200":
          description: Updated OTP config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOtpConfig"
  /admin/blocked-ips:
    get:
      operationId: listBlockedIps
      tags: [admin]
      security:
        - adminBearer: []
      responses:
        "200":
          description: Blocked network addresses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedIpsPage"
    post:
      operationId: blockIp
      tags: [admin]
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IpBlockInput"
      responses:
        "201":
          description: Network IP blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpBlockResult"
  /admin/blocked-ips/{ip}:
    delete:
      operationId: unblockIp
      tags: [admin]
      security:
        - adminBearer: []
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: IP unblocked
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Admin schemas ─────────────────────────────────────────────────────────
...
    AuthProvider:
      type: string
      enum: [phone_otp, clerk]
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResult:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
    AdminStats:
      type: object
      required: [totalUsers, activeUsers, newUsersToday, newUsersThisWeek, trends]
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        newUsersToday:
          type: integer
        newUsersThisWeek:
          type: integer
        trends:
          type: array
          items:
            $ref: "#/components/schemas/AdminTrendDay"
    AdminTrendDay:
      type: object
      required: [date, signups, logins]
      properties:
        date: { type: string, format: date }
        signups: { type: integer }
        logins: { type: integer }
    AdminUser:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - authProvider
        - status
        - isOnline
        - createdAt
        - lastLogin
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        authProvider:
          $ref: "#/components/schemas/AuthProvider"
        status:
          type: string
          enum: [active, suspended]
        isOnline:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: ["string", "null"]
          format: date-time
    AdminUserDetail:
      allOf:
        - $ref: "#/components/schemas/AdminUser"
        - type: object
          required: [loginHistory]
          properties:
            loginHistory:
              type: array
              items:
                $ref: "#/components/schemas/AdminLoginEvent"
    AdminLoginEvent:
      type: object
      required: [id, occurredAt, ip, device, authMethod, source]
      properties:
        id: { type: string }
        occurredAt: { type: string, format: date-time }
        ip:
          type: ["string", "null"]
          description: Unavailable unless verified direct or trusted-proxy policy determines a real client IP. Historical values may reflect shared proxies.
        device: { type: string }
        authMethod: { type: string }
        source: { type: string }
    AdminUserUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, suspended]
    AdminUsersPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    AdminOtpConfig:
      type: object
      required: [enabled, sender, updatedAt, twilio, connectionError]
      properties:
        enabled: { type: boolean }
        sender: { type: string }
        updatedAt:
          type: ["string", "null"]
          format: date-time
        twilio:
          oneOf:
            - $ref: "#/components/schemas/TwilioStatus"
            - type: "null"
        connectionError: { type: ["string", "null"] }
    TwilioStatus:
      type: object
      required: [status, balance, currency, sender, senders]
      properties:
        status: { type: string }
        balance: { type: ["string", "null"] }
        currency: { type: ["string", "null"] }
        sender: { type: ["string", "null"] }
        senders:
          type: array
          items: { type: string }
    AdminOtpConfigInput:
      type: object
      required: [enabled, sender]
      properties:
        enabled: { type: boolean }
        sender: { type: string }
    IpBlockInput:
      type: object
      required: [ip]
      properties:
        ip: { type: string }
        reason: { type: string }
    IpBlockResult:
      type: object
      required: [ip]
      properties:
        ip: { type: string }
    BlockedIp:
      type: object
      required: [ip, reason, createdAt]
      properties:
        ip: { type: string }
        reason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    BlockedIpsPage:
      type: object
      required: [items, policy]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/BlockedIp" }
        policy:
          $ref: "#/components/schemas/ClientIpPolicy"
    ClientIpPolicy:
      type: object
      required: [configured, mode, clientIpAvailable, message]
      properties:
        configured: { type: boolean }
        mode: { type: string, enum: [trusted_proxy, direct, setup_required] }
        clientIpAvailable: { type: boolean }
        message: { type: string }
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT