# lib/api-zod/src/generated/types/adminOtpConfig.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TwilioStatus } from './twilioStatus';

export interface AdminOtpConfig {
  enabled: boolean;
  sender: string;
  /** @nullable */
  updatedAt: Date | null;
  twilio: TwilioStatus | null;
  /** @nullable */
  connectionError: string | null;
}