# lib/api-zod/src/generated/types/authProvider.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type AuthProvider = typeof AuthProvider[keyof typeof AuthProvider];


export const AuthProvider = {
  phone_otp: 'phone_otp',
  clerk: 'clerk',
} as const;