# lib/api-zod/src/generated/types/blockedIp.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlockedIp {
  ip: string;
  /** @nullable */
  reason: string | null;
  createdAt: Date;
}