# lib/api-zod/src/generated/types/blockedIpsPage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlockedIp } from './blockedIp';
import type { ClientIpPolicy } from './clientIpPolicy';

export interface BlockedIpsPage {
  items: BlockedIp[];
  policy: ClientIpPolicy;
}