# lib/api-zod/src/generated/types/twilioStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TwilioStatus {
  status: string;
  /** @nullable */
  balance: string | null;
  /** @nullable */
  currency: string | null;
  /** @nullable */
  sender: string | null;
  senders: string[];
}